+++ apps/ddal-kkak/schemas/blocks/style/index.ts
@@ -1,0 +1,9 @@
+import { z } from "zod";
+import { zodStringNumber } from "@/schemas/blocks";
+
+export const paddingSchema = z.object({
+  paddingTop: zodStringNumber.optional(),
+  paddingBottom: zodStringNumber.optional(),
+  paddingLeft: zodStringNumber.optional(),
+  paddingRight: zodStringNumber.optional(),
+});
